+++ package.json
@@ -3,7 +3,11 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server.js"
+    "serve": "nodemon dist/index.js",
+    "build": "babel src --out-dir dist",
+    "dev": "NODE_ENV=development npm-run-all build serve",
+    "test": "true",
+    "lint": "true"
   },
   "license": "MIT",
   "dependencies": {
@@ -17,13 +21,23 @@
     "mongoose": "^5.9.10",
     "nanoid": "^3.1.3",
     "nodemailer": "^6.4.6",
+    "regenerator-runtime": "^0.13.7",
     "sanitize-html": "^1.23.0",
     "web-push": "^3.4.3"
   },
   "devDependencies": {
+    "@babel/cli": "^7.10.5",
+    "@babel/core": "^7.11.1",
+    "@babel/preset-env": "^7.11.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
-    "nodemon": "^2.0.3"
+    "nodemon": "^2.0.3",
+    "npm-run-all": "^4.1.5"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
   }
 }
