+++ src/database.js
@@ -1,4 +1,9 @@
 const mongoName = '127.0.0.1:27017';
-module.exports = {
+
+const config = {
   url: 'mongodb://' + mongoName + '/sweet'
+};
+
+module.exports = {
+  config,
 };
