+++ src/modules/post/model.js
@@ -1,4 +1,4 @@
-const mongoose = require('mongoose')
+const mongoose = require('mongoose');
 const DBReference = mongoose.Schema.Types.ObjectId
 
 // this is used by older posts instead of the below inlineElementSchema
