+++ src/modules/tag/model.js
@@ -1,4 +1,4 @@
-const mongoose = require('mongoose')
+const mongoose = require('mongoose');
 
 const tagSchema = new mongoose.Schema({
   name: String,
